Update SCCMLab.ps1
Major update. Modified much of the code to use a JSON file for settings, with routines to convert PSObjects/JSON/Hasttables to/from each other to be used by the routines.
Now relies hevaily on splatting for configuring components, however this has trhe side effect of major flexibility by modifying the config file